requirements.txt
-i http://mirrors.aliyun.com/pypi/simple/ --trusted-host mirrors.aliyun.com

redis[hiredis]
asyncpg==0.27.0
celery==5.2.7
importlib-metadata<5.0
fastapi==0.95.0
flower==1.2.0
gunicorn==20.1.0
msgpack==1.0.0
pytest==5.3.5
python-json-logger==0.1.11
ujson==5.7.0
uvicorn==0.21.1
uvloop==0.17.0
sqlalchemy[asyncio]
python-dotenv
typer[all]
aiosql